Improve instrument compilation task
The old version did not take Language.java into account and it did not
trigger compilation on transitive dependencies.

Still did not find a way to safely re-trigger compilation, but a run
with inconsistent state is now aborted and a message is displayed to the
user to manually re-run the last task which will fix the inconsistency
problems.